Extract local site settings to a helper
This will allow us to use locally stored site specific settings from other parts of Yosemite.
